Relocate Rex::Platform:Windows content (fixes MS-1714)
