--- lib/rex/platforms.rb
@@ -1,2 +1,0 @@
-# -*- coding: binary -*-
-require 'rex/platforms/windows'
